ref(symbolication): Introduce stackwalking module
